- Fix sonar reported issue

+++ src/main/java/org/folio/entlinks/service/authority/AuthoritySourceFileService.java
@@ -213,7 +213,7 @@ public class AuthoritySourceFileService {
     }
 
     var sequenceName = existing.getSequenceName();
-    var startNumber = modified.getHridStartNumber();
+    var startNumber = (int) modified.getHridStartNumber();
     var command = String.format("ALTER SEQUENCE %s RESTART WITH %d OWNED BY %s.authority_source_file.sequence_name;",
         sequenceName, startNumber, moduleMetadata.getDBSchemaName(folioExecutionContext.getTenantId()));
     jdbcTemplate.execute(command);
